backend/pyproject.toml
[project]
name = "diff-the-universe"
version = "0.1.0"
description = "Matrix For AI Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "ariadne>=0.26.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "starlette>=0.48.0",
]
